code/requirements.txt
astroid==2.4.2
autopep8==1.5.7
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
greenlet==1.1.1
gunicorn==20.1.0
idna==2.10
isort==5.6.4
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.4.3
mailjet-rest==1.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
neo4j==4.3.4
psycopg2==2.9.1
pycodestyle==2.7.0
pycryptodome==3.10.1
pylint==2.6.0
pyPodcastParser==2.0.0
python-dotenv==0.19.0
pytz==2021.1
requests==2.25.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.4.22
toml==0.10.2
urllib3==1.26.2
Werkzeug==2.0.1
wrapt==1.12.1
